Added Events! Added high res sleep. Moved files

+++ src/Minigin.h
@@ -4,7 +4,7 @@
 #include <vector>
 
 #include "EngineTime.h"
-#include "InputHandler.h"
+#include "Input/InputHandler.h"
 #include "SDL3/SDL_events.h"
 #include "SDL3/SDL_init.h"
 #include "SDL3/SDL_render.h"
@@ -36,7 +36,7 @@ private:
 
 	EngineTime m_engine_time;
 	InputHandler m_input_handler;
-	int m_refresh_rate_delay{};
+	double m_refresh_rate_delay{};
 	SDL_Window* m_window{};
 	SDL_Renderer* m_renderer{};
 	std::vector<std::unique_ptr<GameObject>> m_game_objects;
